+++ cmd/deploy/cmd.go
@@ -4,6 +4,7 @@ import (
 	"net/http"
 
 	"numerous.com/cli/cmd/args"
+	"numerous.com/cli/cmd/errorhandling"
 	"numerous.com/cli/cmd/group"
 	"numerous.com/cli/internal/app"
 	"numerous.com/cli/internal/gql"
@@ -60,8 +61,9 @@ func run(cmd *cobra.Command, args []string) error {
 		Version:    version,
 		Verbose:    verbose,
 	}
+	err := Deploy(cmd.Context(), service, input)
 
-	return Deploy(cmd.Context(), service, input)
+	return errorhandling.ErrorAlreadyPrinted(err)
 }
 
 func init() {