+++ cmd/logout/logout.go
@@ -3,6 +3,7 @@ package logout
 import (
 	"net/http"
 
+	"numerous.com/cli/cmd/errorhandling"
 	"numerous.com/cli/cmd/group"
 	"numerous.com/cli/cmd/output"
 	"numerous.com/cli/internal/auth"
@@ -14,11 +15,10 @@ var LogoutCmd = &cobra.Command{
 	Use:     "logout",
 	Short:   "Logout of the Numerous CLI",
 	GroupID: group.AdditionalCommandsGroupID,
-	RunE:    run,
-}
-
-func run(cmd *cobra.Command, args []string) error {
-	return logout(auth.NumerousTenantAuthenticator)
+	RunE: func(cmd *cobra.Command, args []string) error {
+		err := logout(auth.NumerousTenantAuthenticator)
+		return errorhandling.ErrorAlreadyPrinted(err)
+	},
 }
 
 func logout(a auth.Authenticator) error {