+++ cmd/root.go
@@ -10,7 +10,8 @@ import (
 	"numerous.com/cli/cmd/deploy"
 	"numerous.com/cli/cmd/dev"
 	"numerous.com/cli/cmd/download"
-	"numerous.com/cli/cmd/initialize"
+	"numerous.com/cli/cmd/errorhandling"
+	cmdinit "numerous.com/cli/cmd/init"
 	"numerous.com/cli/cmd/legacy"
 	"numerous.com/cli/cmd/login"
 	"numerous.com/cli/cmd/logout"
@@ -45,8 +46,8 @@ var (
 			"             °°°°°   \n" +
 			"                °°     \n" +
 			"",
-		SilenceUsage:  true,
 		SilenceErrors: true,
+		SilenceUsage:  true,
 		PersistentPreRunE: func(cmd *cobra.Command, args []string) error {
 			output.NotifyFeedbackMaybe()
 
@@ -99,6 +100,11 @@ func commandRequiresAuthentication(invokedCommandName string) bool {
 func Execute() {
 	err := rootCmd.Execute()
 	if err != nil {
+		if !errors.Is(err, errorhandling.ErrAlreadyPrinted) {
+			output.PrintError("Error: %s", "", err.Error())
+			println()
+			rootCmd.Usage() // nolint: errcheck
+		}
 		os.Exit(1)
 	}
 }
@@ -116,7 +122,7 @@ func init() {
 	})
 
 	rootCmd.AddCommand(
-		initialize.InitCmd,
+		cmdinit.InitCmd,
 		login.LoginCmd,
 		logout.LogoutCmd,
 		dev.DevCmd,