+++ python/src/numerous/_client.py
@@ -33,6 +33,7 @@ from numerous.generated.graphql.fragments import (
 from numerous.generated.graphql.input_types import TagInput
 
 
+
 API_URL_NOT_SET = "NUMEROUS_API_URL environment variable is not set"
 MESSAGE_NOT_SET = "NUMEROUS_API_ACCESS_TOKEN environment variable is not set"
 
@@ -41,6 +42,7 @@ class Client:
     def __init__(self, client: GQLClient) -> None:
         self.client = client
         self.organization_id = os.getenv("ORGANIZATION_ID", "default_organization")
+
         auth_token = os.getenv("NUMEROUS_API_ACCESS_TOKEN")
         if not auth_token:
             raise ValueError(MESSAGE_NOT_SET)
@@ -79,6 +81,7 @@ class Client:
         return asyncio.run(
             self._create_collection(collection_key, parent_collection_id)
         )
+
 
     def _create_collection_document_ref(
         self,