+++ python/tests/test_session.py
@@ -2,6 +2,7 @@ from typing import AsyncIterator
 from unittest.mock import Mock
 
 import pytest
+
 from numerous import action, app, container, html, slider
 from numerous.generated.graphql import Client
 from numerous.generated.graphql.all_elements import (